Connected successful pages with sign in and signup
