+++ Frontend/lib/screens/student/student_login.dart
@@ -8,10 +8,12 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:tutorx/screens/student/select_location.dart';
 import 'package:tutorx/screens/student/student_homepage.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:tutorx/screens/common/sign_in_success.dart';
 // import 'package:tutorx/screens/common/map_temp.dart';
 // import 'package:tutorx/screens/student/select_location.dart';
 // import 'package:tutorx/screens/student/student_homepage.dart';
 import 'package:tutorx/utils/auth.dart';
+import 'package:tutorx/utils/shared_preferences_utils.dart';
 import 'package:tutorx/widgets/reusable_widgets.dart';
 import 'package:tutorx/utils/colors.dart';
 import 'package:tutorx/widgets/Google_Sign_In_Button.dart';
@@ -31,18 +33,6 @@ class _StudentSignInState extends State<StudentSignIn> {
   bool _isLoading = false;
   @override
   Widget build(BuildContext context) {
-    //Function to store in Cache
-    Future<void> StoreUserDetailsInCache(String uid) async {
-      var response = await BaseClient().get("/user/$uid").catchError((err) {});
-      var user = usersFromJson(response);
-      // print('Here: ${user.fullname}');
-      final prefs = await SharedPreferences.getInstance();
-      await prefs.setString('fullname', user.fullname);
-      await prefs.setString('uid', user.uid);
-      await prefs.setBool('student', user.student);
-      await prefs.setBool('isLoggedIn', true);
-    }
-
     return Scaffold(
       backgroundColor: Colors.black,
       body: Stack(
@@ -135,10 +125,10 @@ class _StudentSignInState extends State<StudentSignIn> {
                                 var user = jsonDecode(response);
 
                                 if (user["student"] == true) {
-                                  await StoreUserDetailsInCache(uid_temp);
+                                  await SharedPreferencesUtils.StoreUserDetailsInCache(uid_temp);
                                   Navigator.of(context).push(
                                     MaterialPageRoute(
-                                      builder: (context) => StudentHompage(),
+                                      builder: (context) => SignInSuccessful(),
                                     ),
                                   );
                                 } else {
